Fix Readdir and Name()
Statik's Readdir did not behave in the same way as a http.FileSystem's Readdir. Similarly, the Name function in http.FileSystem returned the base name, and not a full file path.

This commit fixes both issues, making both Name and Readdir behave in the same way as in http.Dir()