Fixed map add in loop
Folders were created in a for loop iterating over map keys. This led to an instability in newer golang versions, since adding to a map while looping over it can either produce the key again or not produce it.

In order to successfully generate a folder structure, it was implicitly assumed that each folder has a file, which would not require recursing from a file down to its ancestor folders.

When running without such a requirement, the elements added in the loop sometimes recursed, if their key was produced again in the loop, and sometimes did not. This caused an intermittent issue with tests sometimes failing to find a folder.