Use os.PathSeparator to find files
There was an issue on Windows where the path separator for !windows was hardcoded,
which did not play nice with the filepath.Clean() function.

It replaces all `/` with `\`.
The fix is to use the right path separator which makes it portable.

I don't think it impacts anything, since `/` in URLs will just be resolved
internaly to `\`.